requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appnope==0.1.3
APScheduler==3.10.1
astroid==2.15.5
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
backcall==0.2.0
bitarray==2.7.3
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cytoolz==0.12.1
decorator==5.1.1
dill==0.3.6
dnspython==2.3.0
docstring-to-markdown==0.12
email-validator==2.0.0.post2
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
executing==1.2.0
fastapi==0.95.2
flake8==6.0.0
frozenlist==1.3.3
h11==0.14.0
hexbytes==0.3.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
ipython==8.13.2
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
lazy-object-proxy==1.9.0
lru-dict==1.1.8
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mccabe==0.7.0
multidict==6.0.4
orjson==3.8.14
packaging==23.1
parsimonious==0.9.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.1
pluggy==1.0.0
prompt-toolkit==3.0.38
protobuf==4.23.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pycodestyle==2.10.0
pycryptodome==3.18.0
pydantic==1.10.8
pydocstyle==6.3.0
pyflakes==3.0.1
Pygments==2.15.1
pylint==2.17.4
pymongo==4.3.3
pyrsistent==0.19.3
python-dotenv==1.0.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.3
python-multipart==0.0.6
python-telegram-bot==20.3
pytoolconfig==1.2.5
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
rlp==3.0.0
rope==1.8.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
SQLAlchemy==2.0.15
stack-data==0.6.2
starlette==0.27.0
tomli==2.0.1
tomlkit==0.11.8
toolz==0.12.0
traitlets==5.9.0
typing_extensions==4.6.2
tzlocal==5.0.1
ujson==5.7.0
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
web3==6.4.0
websockets==11.0.3
whatthepatch==1.0.5
wrapt==1.15.0
yapf==0.33.0
yarl==1.9.2
